Fixs to make single point float build work

+++ include/libsurvive/survive_optimizer.h
@@ -32,7 +32,7 @@ typedef struct {
 	FLT current_bias;
 	SurvivePose initialPose;
 
-	double *parameters;
+	FLT *parameters;
 	struct mp_par_struct *parameters_info;
 
 	int poseLength;
@@ -90,7 +90,7 @@ SURVIVE_EXPORT BaseStationCal *survive_optimizer_get_calibration(survive_optimiz
 
 SURVIVE_EXPORT int survive_optimizer_get_sensors_index(const survive_optimizer *ctx);
 
-SURVIVE_EXPORT double *survive_optimizer_get_sensors(survive_optimizer *ctx);
+SURVIVE_EXPORT FLT *survive_optimizer_get_sensors(survive_optimizer *ctx);
 
 SURVIVE_EXPORT void survive_optimizer_setup_pose(survive_optimizer *mpfit_ctx, const SurvivePose *pose, bool isFixed,
 												 int use_jacobian_function);
@@ -118,8 +118,8 @@ SURVIVE_EXPORT mp_config *survive_optimizer_precise_config();
 
 SURVIVE_EXPORT int survive_optimizer_nonfixed_cnt(const survive_optimizer *optimizer);
 
-SURVIVE_EXPORT void survive_optimizer_get_nonfixed(const survive_optimizer *optimizer, double *params);
-SURVIVE_EXPORT void survive_optimizer_set_nonfixed(survive_optimizer *optimizer, double *params);
+SURVIVE_EXPORT void survive_optimizer_get_nonfixed(const survive_optimizer *optimizer, FLT *params);
+SURVIVE_EXPORT void survive_optimizer_set_nonfixed(survive_optimizer *optimizer, FLT *params);
 
 #ifdef __cplusplus
 }
